+++ AtomF.m
@@ -12,7 +12,7 @@ classdef AtomF < Atom
     methods (Access = protected)
         
         function SetAtomicParameters(obj)
-            obj.atomType = EnumAtom.F;
+            obj.atomType = uint8(EnumAtom.F);
             obj.atomicMass = 18.9984032*Arguments.GetInstance().GetGMolin2AU();
             obj.coreCharge = 7.0;
             obj.numberValenceElectrons = 7;
